t0lte: fix RIL
The KK ks crashed when setting ril.ks.status property, we work around
this by using the D2 ks (which does not set this property) and wrapping it
with a bash script because D2's ks does not support some flags passed
by T0's qcks. We use the bash script to set the ril.ks.status property so qcks
knows it can continue.

Change-Id: Icc1a64b93edc38bc477acc953595280929a46549