Allow NavBar Menu in Settings
Allow t0lte to enable the NavBar Preferences in Settings. Currently these are hidden from t0lte users. Not sure if this change is correct. I don't want to enable the navbar by default, but give users the options by having it disabled by default, but enable the preferences for the NavBar. Similar to how Samsung Galaxy S4 devices have the NavBar disabled by default, but allow users to easily change a preference to enable it.